requirements.txt
asgiref==3.7.2
Django==4.2.2
django-js-asset==2.1.0
django-mptt==0.14.0
Pillow==10.0.0
python-dotenv==1.0.0
sqlparse==0.4.4
